Rename lock.h to pylock.h
